main: Move shutdown to own file
